requirement.txt
wikipedia@1.4.0
speechrecognition@3.14.3
pywin32@310
PyAudio@0.2.14
markdown2
google-genai